Move FakeServiceHandler to -testing so that it can be used in other places, too.
